saving module name changes
