+++ screen_control_bot.py
@@ -3,6 +3,12 @@ import time
 import winsound
 import pyautogui
 import ocr
+
+
+pyautogui.PAUSE = 0
+pyautogui.FAILSAFE = True
+MAX_FAILED_MOVE_ATTEMPTS = 15
+FAILED_MOVE_ATTEMPTS = 0
 FAILED_MOVE_ATTEMPTS = 0
 MAX_FAILED_MOVE_ATTEMPTS = 0
 class ScreenBot():
@@ -175,3 +181,49 @@ class FailedMoveAttempt(Exception):
     "Raised when there's too many failed movement attempts"
     pass
 
+
+def randomBrowsing():
+    screenWidth, screenHeight = pyautogui.size()
+    x = random.randint(1,screenWidth)
+    y = random.randint(1,screenHeight)
+    moveTime = random.randrange(15, 30, 1) / 10
+    human_move(x, y, moveTime, 4)
+
+
+def flick_pray(rapidHeal):
+    randInterval = random.randint(20,30)/100
+    pyautogui.press('f5', interval=randInterval)
+    time.sleep(random.randint(25, 35) / 100)
+    randInterval = random.normalvariate(25, 3) / 10
+    human_move(rapidHeal[0], rapidHeal[1], randInterval, 2)
+    doClick(clicks = 1, duration = (random.normalvariate(25, 3)/100))
+    time.sleep(random.randint(150, 200) / 100)
+    doClick(clicks=1, duration=(random.normalvariate(25, 3) / 100))
+    time.sleep(random.normalvariate(175,15)/100)
+    randInterval = random.normalvariate(25, 3) / 100
+    pyautogui.press('esc', interval=randInterval)
+
+
+def checkAndConsume(consumedType,consumedTimer,coordsList,consumedList,timeElapsed):
+    if timeElapsed - consumedTimer[consumedType] * consumedList[consumedType] > consumedTimer[consumedType]:
+        typename = ["Overload","Prayer Potion"]
+        print(("Drinking " + str(typename[consumedType])+"\n")*3)
+        time.sleep(2)
+        #move to our potion
+        x = coordsList[consumedType][0][0] +random.normalvariate(0,2)
+        y = coordsList[consumedType][0][1] +random.normalvariate(0,2)
+        moveTime = random.randrange(20,40,1)/10
+        human_move(x, y, moveTime, 3)
+        #click
+        doClick(duration=random.normalvariate(20,3)/100)
+        time.sleep(.5)
+        #Move to a random location occasionly
+        if random.randint(0,20) > 19:
+            randomBrowsing()
+        #update our consumption
+        consumedList[consumedType] += 1
+        if consumedList[consumedType] % 4 == 0:
+            coordsList[consumedType].pop(0)
+        if typename[consumedType]=="Overload" and len(coordsList[consumedType])==1 and consumedList[consumedType] % 4 == 3:
+            coordsList[consumedType].pop(0)
+            consumedList[consumedType] += 1
